% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_mean.R
\name{find_mean}
\alias{find_mean}
\title{A Function to find the mean of a subject's scores}
\usage{
find_mean(scores)
}
\arguments{
\item{scores}{an R object containing scores for each response}
}
\value{
the value of the mean that was inserted into outputlocation
}
\description{
A Function to find the mean of a subject's scores
}

